ai_detection/requirements.txt
# AI Detection Dependencies
# Separate from main SusScrofa requirements for modularity

# PyTorch (CUDA 12.4 - adjust version if needed)
--extra-index-url https://download.pytorch.org/whl/cu124
torch>=2.0.0
torchvision>=0.15.0

# Core dependencies for SPAI
opencv-python>=4.10.0
pyyaml>=6.0.1
scipy>=1.14.0
timm==0.4.12
yacs>=0.1.8
numpy>=1.26.4
torchmetrics>=1.4.0
tqdm>=4.66.4
pillow>=10.4.0
einops>=0.8.0
ftfy>=6.1.0
regex>=2023.0.0

# CLIP for alternative backbones (optional but used in backbones.py)
git+https://github.com/openai/CLIP.git

# HuggingFace Transformers (for Swin-based detectors: Organika/sdxl-detector, etc.)
transformers>=4.36.0
safetensors>=0.4.0

# Optional: For additional features
# tensorboard  # For training visualization (not needed for inference)
# neptune      # For experiment tracking (not needed for inference)
